Panels, properies, and further file migration.
